Modularize data layer
