+++ app/build.gradle
@@ -63,6 +63,7 @@ android {
 
 dependencies {
     implementation project(":domain")
+    implementation project(":data")
 
     implementation fileTree(dir: "libs", include: ["*.jar"])
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
