+++ app/src/main/java/com/maximo/douglas/githubconsumer/ui/gitpullrequests/SelectGitPullRequestViewModel.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.maximo.douglas.domain.entities.gitpullrequest.GitPullRequest
-import com.maximo.douglas.githubconsumer.service.gitpullrequest.GitPullRequestRemoteDataSource
+import com.maximo.douglas.data.remote.gitpullrequest.GitPullRequestRemoteDataSource
 import com.maximo.douglas.githubconsumer.utils.CoroutineContextProvider
 import kotlinx.coroutines.CoroutineExceptionHandler
 import kotlinx.coroutines.launch
@@ -13,7 +13,7 @@ import kotlinx.coroutines.withContext
 
 @Suppress("UNCHECKED_CAST")
 class SelectGitPullRequestViewModel(
-    private val gitPullRequestRemoteDataSource: GitPullRequestRemoteDataSource,
+    private val gitPullRequestRemoteDataSource: com.maximo.douglas.data.remote.gitpullrequest.GitPullRequestRemoteDataSource,
     private val contextProvider: CoroutineContextProvider
 ) : ViewModel() {
