+++ app/src/test/java/com/maximo/douglas/githubconsumer/viewmodels/gitpullrequest/context/SelectGitPullRequestViewModelTest.kt
@@ -3,7 +3,7 @@ package com.maximo.douglas.githubconsumer.viewmodels.gitpullrequest.context
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import androidx.lifecycle.Observer
 import com.maximo.douglas.domain.entities.gitpullrequest.GitPullRequest
-import com.maximo.douglas.githubconsumer.service.gitpullrequest.GitPullRequestRemoteDataSource
+import com.maximo.douglas.data.remote.gitpullrequest.GitPullRequestRemoteDataSource
 import com.maximo.douglas.githubconsumer.ui.gitpullrequests.SelectGitPullRequestViewModel
 import com.maximo.douglas.githubconsumer.ui.gitpullrequests.SelectGitPullRequestViewModel.SelectGitPullRequestViewModelState
 import com.maximo.douglas.githubconsumer.viewmodels.utils.TestContextProvider
@@ -27,7 +27,7 @@ open class SelectGitPullRequestViewModelTest {
     val testCoroutineRule = TestCoroutineRule()
 
     @Mock
-    private lateinit var mGitPullRequestRemoteDataSource: GitPullRequestRemoteDataSource
+    private lateinit var mGitPullRequestRemoteDataSource: com.maximo.douglas.data.remote.gitpullrequest.GitPullRequestRemoteDataSource
 
     @Mock
     private lateinit var mViewStateObserver: Observer<SelectGitPullRequestViewModelState>
