+++ data/src/main/java/com/maximo/douglas/data/remote/gitpullrequest/GitPullRequestRemoteDataSourceImpl.kt
@@ -1,4 +1,4 @@
-package com.maximo.douglas.githubconsumer.service.gitpullrequest
+package com.maximo.douglas.data.remote.gitpullrequest
 
 class GitPullRequestRemoteDataSourceImpl(
     private val mGitPullRequestApiDataSource: GitPullRequestApiDataSource
