+++ data/src/main/java/com/maximo/douglas/data/remote/gitrepository/GitRepositoryApiDataSource.kt
@@ -1,4 +1,4 @@
-package com.maximo.douglas.githubconsumer.service.gitrepository
+package com.maximo.douglas.data.remote.gitrepository
 
 import com.maximo.douglas.domain.entities.gitrepository.GitRepositoryApiResponse
 import retrofit2.http.GET
