+++ settings.gradle
@@ -1,3 +1,4 @@
+include ':data'
 include ':domain'
 include ':app'
 rootProject.name = "github-api-consumer"
